refactor(game): give game.rs a directory, and its fixtures a home
`commands/game.rs` is 4358 lines, over half of them tests, and it is the
largest file in the backend. Decomposing it is the point of the work that
follows; this commit only opens the door.

Two things happen here and nothing else. `game.rs` becomes `game/mod.rs`,
which git records as a rename. And the eight fixtures that several unrelated
clusters of tests share — `make_bootstrap_test_game`, `sample_stats_state`,
the two imported-world builders, and four smaller ones — move to
`game/testkit.rs`.

The fixtures have to move first, before any production code does. They are
what makes the file's tests inseparable: the startup tests and the bootstrap
tests have almost nothing to do with each other, but both call
`sample_stats_state`. Peel a cluster out without a shared home for those, and
the choice is duplicating a fixture or having one submodule's tests reach
sideways into another's. `testkit` is reachable as
`crate::commands::game::testkit::*` — an absolute path, so it survives a
submodule being nested further later.

Visibility stays as narrow as it was: `mod testkit;` is private to `game`,
and the fixtures are `pub(super)`, which Rust makes visible to `game` and its
descendants. That is exactly the set of modules that will need them.

No behaviour changes and no test changes. The check that says so is the test
list: 212 lib tests before, 212 after, identical name for name. That check is
worth naming, because the failure it guards against is quiet — a submodule
whose `mod tests;` declaration is forgotten compiles clean, reports fewer
passing tests, and still looks green.
